requirements.txt
apache-airflow==2.7.2
pyspark==3.4.1
psycopg2-binary==2.9.7
pandas==2.0.3
numpy==1.24.3
sqlalchemy==1.4.49
redis==4.6.0
pytz==2023.3
requests==2.31.0
cryptography==41.0.4
